0.0.1/refactoring: fix newline(dos -> unix)
