--- lib/sf_cli/sf/operations.rb
@@ -1,4 +1,0 @@
-require_relative './org'
-require_relative './data'
-require_relative './sobject'
-require_relative './project'
