+++ lib/sf_cli/sf/sobject.rb
@@ -1,35 +1,38 @@
-require 'json'
-require_relative './base'
-
 module SfCli
   class Sf
-    # ==== description
-    # The class representing *sf* *sobject*
-    #
-    # command reference: https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_sobject_commands_unified.htm
-    #
-    class SObject < Base
+    class Sobject
+      def initialize(_sf)
+        @sf  = _sf
+      end
 
-      # get the schema information of a Salesforce object and returns a hash object. It's equivalent to use *sf* *sobject* *describe*
-      #
-      # *objectType* --- object type (ex: Account)
-      #
-      def describe(objectType)
-        json = JSON.parse `sf sobject describe --json --sobject #{objectType} #{flag :"target-org", target_org}  #{null_stderr_redirection}`
-        raise StandardError.new(%|sf sobject describe: failed. (sobject: #{objectType})|) if json['status'] != 0
-
+      def describe(object_type, target_org: nil)
+        flags    = {
+          :"sobject"    => object_type,
+          :"target-org" => target_org,
+        }
+        switches = [:json]
+        json = sf.exec(category, __method__, flags: flags, switches: switches, redirection: :null_stderr)
         json['result']
       end
 
-      # returns a list of Salesforce object name (object's API name). It's equivalent to *sf* *sobject* *list*
-      #
-      # *object_type* --- all or custom (default: all)
-      #
-      def list(object_type: 'all')
-        json = JSON.parse `sf sobject list --json --sobject #{object_type} #{flag :"target-org", target_org} #{null_stderr_redirection}`
-        raise StandardError.new(%|sf sobject list: failed.|) if json['status'] != 0
-
+      def list(object_type, target_org: nil)
+        flags    = {
+          :"sobject"    => (object_type.to_sym == :custom ? :custom : :all),
+          :"target-org" => target_org,
+        }
+        switches = [:json]
+        json = sf.exec(category, __method__, flags: flags, switches: switches, redirection: :null_stderr)
         json['result']
+      end
+
+      private
+
+      def category
+        self.class.name.split('::').last.downcase
+      end
+
+      def sf
+        @sf
       end
     end
   end
